requirements.txt
Flask==1.0.2
psycopg2-binary==2.8.6
requests==2.19.1
python-dotenv==0.13.0
bcrypt==3.1.7
